refactor: 배럴 패턴으로 변경
