+++ src/__tests__/Validator.test.js
@@ -1,5 +1,5 @@
 import { describe, test, expect } from 'vitest';
-import Validator from '../utils/validators/Validator.js';
+import Validator from '../utils/validators/Validator';
 
 describe('Validator >', () => {
   test('값이 없으면 오류가 발생한다.', () => {
